add FirehoseInstance file
