+++ lib/log.js
@@ -1,20 +1,17 @@
 'use strict';
 
 const EventEmmiter = require('events');
-
-const { STS, Firehose } = require('./aws-wrappers');
 const { arrayChunk } = require('./helpers/utils');
-
 const Validator = require('./helpers/validator');
 const LogError = require('./log-error');
+const serverlessConfiguration = require('./serverless-configuration');
+const FirehoseInstance = require('./firehose-instance');
 
-const ARN_DURATION = 1800; // 30 min
+
 const MAX_ATTEMPTS = 3;
-const MAX_TIMEOUT = 500;
 const DELIVERY_STREAM_PREFIX = 'JanisTraceFirehose';
 const LOGS_BATCH_LIMIT = 500;
 
-const sts = new STS();
 
 const emitter = new EventEmmiter();
 
@@ -119,7 +116,7 @@ class Log {
 
 		try {
 
-			const firehose = await this.getFirehoseInstance();
+			const firehose = await FirehoseInstance.getFirehoseInstance();
 
 			await Promise.all(
 				logsBatches.map(logs => firehose.putRecordBatch(
@@ -145,51 +142,15 @@ class Log {
 		}
 	}
 
-	static async getFirehoseInstance() {
-
-		if(!this.validCredentials()) {
-
-			const firehoseParams = {
-				region: process.env.AWS_DEFAULT_REGION,
-				httpOptions: { timeout: MAX_TIMEOUT }
-			};
-
-			if(this.roleArn) {
-				firehoseParams.credentials = await this.getCredentials();
-				this.credentialsExpiration = new Date(firehoseParams.credentials.expiration);
-			}
-
-			this.firehose = new Firehose(firehoseParams);
-		}
-
-		return this.firehose;
-	}
-
-	static validCredentials() {
-		return this.firehose
-			&& this.credentialsExpiration
-			&& this.credentialsExpiration >= new Date();
-	}
-
-	static async getCredentials() {
-
-		const assumedRole = await sts.assumeRole({
-			RoleArn: this.roleArn,
-			RoleSessionName: this.serviceName,
-			DurationSeconds: ARN_DURATION
-		});
-
-		if(!assumedRole)
-			throw new LogError('Failed to assume role, invalid response.', LogError.codes.ASSUME_ROLE_ERROR);
-
-		const { Credentials, Expiration } = assumedRole;
-
-		return {
-			accessKeyId: Credentials.AccessKeyId,
-			secretAccessKey: Credentials.SecretAccessKey,
-			sessionToken: Credentials.SessionToken,
-			expiration: Expiration
-		};
+	/**
+	 * Returns the sls helpers needed for serverles configuration.
+	 *
+	 * @readonly
+	 * @static
+	 * @memberof Log
+	 */
+	static get serverlessConfiguration() {
+		return serverlessConfiguration();
 	}
 }
