+++ lib/serverless-configuration.js
@@ -6,10 +6,10 @@ module.exports = () => {
 
 	const logRoleArn = Settings.get('logRoleArn');
 
-	return [
+	return [[
 		'iamStatement', {
 			action: 'Sts:AssumeRole',
 			resource: logRoleArn
 		}
-	];
+	]];
 };
