--- tests/serverless-configuration.js
@@ -1,1 +1,0 @@
-// to implement
